# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "MeetingCreate": {
        "default": null,
        "properties": {
          "attendees": {
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "meeting-note",
              "wireframe"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "postedBy": {
            "type": "string"
          },
          "source": {
            "enum": [
              "meeting",
              "feedback",
              "suggestion"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "date",
          "title"
        ],
        "type": "object"
      },
      "MeetingUpdate": {
        "default": null,
        "description": "Only the fields present are changed.",
        "properties": {
          "attendees": {
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "kind": {
            "enum": [
              "meeting-note",
              "wireframe"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "postedBy": {
            "type": "string"
          },
          "productId": {
            "default": "",
            "description": "Id of the product to build into; empty string clears it.",
            "type": "string"
          },
          "source": {
            "enum": [
              "meeting",
              "feedback",
              "suggestion"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          }
        },
        "type": "object"
      },
      "MeetingView": {
        "default": null,
        "properties": {
          "attendees": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "date": {
            "format": "date",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "meeting-note",
              "wireframe"
            ],
            "type": "string"
          },
          "movedTo": {
            "$ref": "#/components/schemas/ProjectRef",
            "default": "",
            "description": "The product they have actually been sent to. A fact."
          },
          "notes": {
            "type": "string"
          },
          "postedBy": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/ProjectRef",
            "default": "",
            "description": "The product these screens are built to fit. A plan."
          },
          "projectId": {
            "type": "string"
          },
          "screens": {
            "items": {
              "$ref": "#/components/schemas/ScreenView"
            },
            "type": "array"
          },
          "source": {
            "enum": [
              "meeting",
              "feedback",
              "suggestion"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProjectCreate": {
        "default": null,
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "customer": {
            "type": "string"
          },
          "id": {
            "default": "",
            "description": "Supplied when the client already named it; generated otherwise.",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ProjectRef": {
        "default": null,
        "description": "Just enough of a project to name it.",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProjectUpdate": {
        "default": null,
        "description": "Only the fields present are changed.",
        "properties": {
          "accent": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "customer": {
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "solution": {
            "type": "string"
          },
          "spend": {
            "type": "number"
          },
          "stage": {
            "type": "string"
          },
          "statusLabel": {
            "type": "string"
          },
          "statusTone": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProjectView": {
        "default": null,
        "description": "A project of either kind; archived means Customer.",
        "properties": {
          "accent": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "customer": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "solution": {
            "type": "string"
          },
          "spend": {
            "type": "number"
          },
          "stage": {
            "type": "string"
          },
          "statusLabel": {
            "type": "string"
          },
          "statusTone": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ScreenHtml": {
        "default": null,
        "properties": {
          "html": {
            "type": "string"
          }
        },
        "required": [
          "html"
        ],
        "type": "object"
      },
      "ScreenView": {
        "default": null,
        "description": "A screen without its page. The html is fetched per screen, because a meeting's pages together run to megabytes.",
        "properties": {
          "change": {
            "default": "",
            "description": "`new` when the product has never had it, `modified` when it changed since it was sent, null when they agree.",
            "enum": [
              "new",
              "modified"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "movedAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "platform": {
            "enum": [
              "PC",
              "Mobile"
            ],
            "type": "string"
          },
          "position": {
            "format": "int32",
            "type": "integer"
          },
          "screenType": {
            "enum": [
              "Screen",
              "Popup",
              "Drawer"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ScreenWrite": {
        "default": null,
        "description": "One generated screen. `parentId` may name another screen in the same request; parents need not be listed first.",
        "properties": {
          "html": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "platform": {
            "enum": [
              "PC",
              "Mobile"
            ],
            "type": "string"
          },
          "screenType": {
            "enum": [
              "Screen",
              "Popup",
              "Drawer"
            ],
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ScreensMoved": {
        "default": null,
        "description": "Records that the set has been sent to a product.",
        "properties": {
          "projectId": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "projectId"
        ],
        "type": "object"
      },
      "ScreensReplace": {
        "default": null,
        "description": "The whole set after a generation. Screens matched by name keep their id, their placement and their moved mark; the rest are new.",
        "properties": {
          "screens": {
            "items": {
              "$ref": "#/components/schemas/ScreenWrite"
            },
            "type": "array"
          }
        },
        "required": [
          "screens"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Projects, meetings, generated screens, and the requests they become.",
    "license": {
      "name": "Proprietary"
    },
    "title": "WE-ADK API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meetings/{id}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Meetings"
        ]
      },
      "get": {
        "operationId": "get_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Meetings"
        ]
      },
      "patch": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Meetings"
        ]
      }
    },
    "/api/meetings/{id}/screens": {
      "delete": {
        "operationId": "reset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Reset: the screens go, the notes stay.",
        "tags": [
          "Meetings"
        ]
      },
      "put": {
        "operationId": "replaceScreens",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreensReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Replace the set after a generation; unmentioned screens are kept as they are.",
        "tags": [
          "Meetings"
        ]
      }
    },
    "/api/meetings/{id}/screens/moved": {
      "post": {
        "operationId": "markMoved",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreensMoved"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Record that the whole set has been sent to a product.",
        "tags": [
          "Meetings"
        ]
      }
    },
    "/api/projects": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List projects; omit `archived` for all of them.",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{projectId}/meetings": {
      "get": {
        "operationId": "list_2",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeetingView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "Meetings"
        ]
      },
      "post": {
        "operationId": "create_1",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingView"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "Meetings"
        ]
      }
    },
...
    "/api/screens/{id}/html": {
      "get": {
        "operationId": "html",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenHtml"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "The page itself, fetched per screen rather than with the meeting.",
        "tags": [
          "Meetings"
        ]
      },
      "put": {
        "operationId": "saveHtml",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenHtml"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenView"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Save the page. Writing it is what marks the screen modified.",
        "tags": [
          "Meetings"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}